@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.0",
     "ext-phalcon": ">=3.0",
     "phalcon/incubator": "~3.0"
   },
@@ -19,7 +19,7 @@
     "phalcon/devtools": "dev-master",
     "phpunit/phpunit": "~5.0",
     "mockery/mockery": "~0.9.4",
-    "satooshi/php-coveralls": "~0.6"
+    "satooshi/php-coveralls": "~1.0"
   },
   "autoload": {
     "psr-4": {
