@@ -2,6 +2,8 @@
 
 namespace Luxury\Middleware;
 
+use Phalcon\Events\Event;
+
 /**
  * Interface AfterMiddleware
  *
@@ -14,12 +16,12 @@ interface AfterMiddleware
     /**
      * Called after the execution of handler
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function after($event, $source, $data = null);
+    public function after(Event $event, $source, $data = null);
 }
