@@ -2,6 +2,8 @@
 
 namespace Luxury\Middleware;
 
+use Phalcon\Events\Event;
+
 /**
  * Interface InitMiddleware
  *
@@ -14,12 +16,12 @@ interface InitMiddleware
     /**
      * Called on the initialization
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function init($event, $source, $data = null);
+    public function init(Event $event, $source, $data = null);
 }
