@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,13 +10,19 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Annotations implements Providable
+class Annotations extends Provider
 {
+    protected $name = Services::ANNOTATIONS;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Phalcon\Annotations\Adapter\Memory
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::ANNOTATIONS, \Phalcon\Annotations\Adapter\Memory::class);
+        return new \Phalcon\Annotations\Adapter\Memory;
     }
 }
