@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,18 +10,20 @@ use Phalcon\DiInterface;
  *
  * @package     Luxury\Providers
  */
-class Auth implements Providable
+class Auth extends Provider
 {
+
+    protected $name = Services::AUTH;
+
+    protected $shared = true;
 
     /**
      * @param \Phalcon\DiInterface $di
      *
-     * @return void
+     * @return \Luxury\Auth\AuthManager
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::AUTH, function () {
-            return new \Luxury\Auth\AuthManager();
-        });
+        return new \Luxury\Auth\AuthManager();
     }
 }
