@@ -4,6 +4,7 @@ namespace Luxury\Providers\Cli;
 
 use Luxury\Constants\Services;
 use Luxury\Interfaces\Providable;
+use Luxury\Providers\Provider;
 use Phalcon\DiInterface;
 use Phalcon\Security;
 
@@ -12,21 +13,25 @@ use Phalcon\Security;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Dispatcher implements Providable
+class Dispatcher extends Provider
 {
+    protected $name = Services::DISPATCHER;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(
-            Services::DISPATCHER,
-            function () {
+        //$di->setShared(
+            //Services::DISPATCHER,
+            //function () {
                 /* @var \Phalcon\Di $this */
                 $dispatcher = new \Phalcon\Cli\Dispatcher();
 
                 // Create an events manager
-                $eventsManager = $this->getShared(Services::EVENTS_MANAGER);
+                $eventsManager = $di->getShared(Services::EVENTS_MANAGER);
 
                 // Assign the events manager to the dispatcher
                 $dispatcher->setEventsManager($eventsManager);
@@ -34,7 +39,7 @@ class Dispatcher implements Providable
                 $dispatcher->setDefaultNamespace('\App\Cli\Tasks');
 
                 return $dispatcher;
-            }
-        );
+            //}
+        //);
     }
 }
