@@ -11,13 +11,20 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Cookies implements Providable
+class Cookies extends Provider
 {
+
+    protected $name = Services::COOKIES;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Phalcon\Http\Response\Cookies
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::COOKIES, \Phalcon\Http\Response\Cookies::class);
+        return new \Phalcon\Http\Response\Cookies;
     }
 }
