@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,24 +10,28 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Foundation\Bootstrap
  */
-class Database implements Providable
+class Database extends Provider
 {
+    protected $name = Services::DB;
+
+    protected $shared = true;
+
     /**
      * Database connection is created based in the parameters defined in the configuration file
      *
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Phalcon\Db\Adapter\Pdo
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::DB, function () {
-            /* @var \Phalcon\Di $this */
-            $dbConfig = $this->getShared(Services::CONFIG)->database->toArray();
-            $adapter  = $dbConfig['adapter'];
-            unset($dbConfig['adapter']);
+        $dbConfig = $di->getShared(Services::CONFIG)->database->toArray();
 
-            $class = 'Phalcon\Db\Adapter\Pdo\\' . $adapter;
+        $adapter = $dbConfig['adapter'];
+        unset($dbConfig['adapter']);
 
-            return new $class($dbConfig);
-        });
+        $class = 'Phalcon\Db\Adapter\Pdo\\' . $adapter;
+
+        return new $class($dbConfig);
     }
 }
