@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,13 +10,19 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Escaper implements Providable
+class Escaper extends Provider
 {
+    protected $name = Services::ESCAPER;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Phalcon\Escaper
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::ESCAPER, \Phalcon\Escaper::class);
+        return new \Phalcon\Escaper;
     }
 }
