@@ -11,13 +11,19 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Filter implements Providable
+class Filter extends Provider
 {
+    protected $name = Services::FILTER;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return mixed|\Phalcon\Filter
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::FILTER, \Phalcon\Filter::class);
+        return new \Phalcon\Filter;
     }
 }
