@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,22 +10,24 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Foundation\Bootstrap
  */
-class Flash implements Providable
+class Flash extends Provider
 {
+    protected $name = Services::FLASH;
+
+    protected $shared = false;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Phalcon\Flash\Direct
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->set(Services::FLASH, function () {
-            return new \Phalcon\Flash\Direct([
-                'error'   => 'alert alert-danger',
-                'success' => 'alert alert-success',
-                'notice'  => 'alert alert-info',
-                'warning' => 'alert alert-warning'
-            ]);
-        });
-
-        $di->setShared(Services::FLASH_SESSION, \Phalcon\Flash\Session::class);
+        return new \Phalcon\Flash\Direct([
+            'error'   => 'alert alert-danger',
+            'success' => 'alert alert-success',
+            'notice'  => 'alert alert-info',
+            'warning' => 'alert alert-warning'
+        ]);
     }
 }
