@@ -4,6 +4,7 @@ namespace Luxury\Providers\Http;
 
 use Luxury\Constants\Services;
 use Luxury\Interfaces\Providable;
+use Luxury\Providers\Provider;
 use Phalcon\DiInterface;
 
 /**
@@ -11,33 +12,37 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Bootstrap\Services
  */
-class Dispatcher implements Providable
+class Dispatcher extends Provider
 {
+    protected $name = Services::DISPATCHER;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(
+        /*$di->setShared(
             Services::DISPATCHER,
-            function () {
+            function () {*/
                 /* @var \Phalcon\Di $this */
                 $dispatcher = new \Phalcon\Mvc\Dispatcher();
 
                 // Create an events manager
-                $eventsManager = $this->getShared(Services::EVENTS_MANAGER);
+                $eventsManager = $di->getShared(Services::EVENTS_MANAGER);
 
                 // Listen for events produced in the dispatcher using the Security plugin
                 $eventsManager->attach(
                     'dispatch:beforeExecuteRoute',
-                    $this->getShared(Services::SECURITY)
+                    $di->getShared(Services::SECURITY)
                 );
 
                 // Assign the events manager to the dispatcher
                 $dispatcher->setEventsManager($eventsManager);
 
                 return $dispatcher;
-            }
-        );
+            /*}
+        );*/
     }
 }
