@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -11,13 +10,19 @@ use Phalcon\DiInterface;
  *
  * @package     Luxury\Providers
  */
-class HttpClient implements Providable
+class HttpClient extends Provider
 {
+    protected $name = Services::HTTP_CLIENT;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
+     *
+     * @return \Luxury\Http\Client
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::HTTP_CLIENT, \Luxury\Http\Client::class);
+        return new \Luxury\Http\Client;
     }
 }
