@@ -4,7 +4,6 @@ namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
 use Luxury\Exceptions\SessionAdapterNotFound;
-use Luxury\Interfaces\Providable;
 use Phalcon\DiInterface;
 
 /**
@@ -12,32 +11,35 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Foundation\Bootstrap
  */
-class Session implements Providable
+class Session extends Provider
 {
+    protected $name = Services::SESSION;
+
+    protected $shared = true;
 
     /**
      * Start the session the first time some component request the session service
      *
      * @param \Phalcon\DiInterface $di
+     *
+     * @throws \Luxury\Exceptions\SessionAdapterNotFound
+     *
+     * @return mixed|\Phalcon\Session\Adapter|\Phalcon\Session\AdapterInterface
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
-        $di->setShared(Services::SESSION, function () {
-            /* @var \Phalcon\Di $this */
-            /* @var \Phalcon\Session\Adapter|\Phalcon\Session\AdapterInterface $session */
-            $class =
-                'Phalcon\Session\Adapter\\' . $this->getShared(Services::CONFIG)->session->adapter;
-            try {
-                $session = new $class();
-            } catch (\Exception $e) {
-                throw new SessionAdapterNotFound($e);
-            }
-
-            $session->start();
-
-            return $session;
-        });
-
         $di->set(Services::SESSION_BAG, \Phalcon\Session\Bag::class);
+
+        /* @var \Phalcon\Session\Adapter|\Phalcon\Session\AdapterInterface $session */
+        $class = 'Phalcon\Session\Adapter\\' . $di->getShared(Services::CONFIG)->session->adapter;
+        try {
+            $session = new $class();
+        } catch (\Exception $e) {
+            throw new SessionAdapterNotFound($e);
+        }
+
+        $session->start();
+
+        return $session;
     }
 }
