@@ -3,7 +3,6 @@
 namespace Luxury\Providers;
 
 use Luxury\Constants\Services;
-use Luxury\Interfaces\Providable;
 use Luxury\View\Engine\Extensions\PhpFunction as PhpFunctionExtension;
 use Phalcon\DiInterface;
 
@@ -12,39 +11,42 @@ use Phalcon\DiInterface;
  *
  * @package Luxury\Foundation\Bootstrap
  */
-class View implements Providable
+class View extends Provider
 {
+    protected $name = Services::VIEW;
+
+    protected $shared = true;
+
     /**
      * @param \Phalcon\DiInterface $di
      */
-    public function register(DiInterface $di)
+    protected function register(DiInterface $di)
     {
         $di->setShared(Services::TAG, \Phalcon\Tag::class);
         $di->setShared(Services::ASSETS, \Phalcon\Assets\Manager::class);
-        $di->setShared(Services::VIEW, function () {
-            /* @var \Phalcon\Di $this */
+        //$di->setShared(Services::VIEW, function () {
 
-            $view = new \Phalcon\Mvc\View();
+        $view = new \Phalcon\Mvc\View();
 
-            $view->setViewsDir($this->getShared(Services::CONFIG)->view->viewsDir);
+        $view->setViewsDir($di->getShared(Services::CONFIG)->view->viewsDir);
 
-            $view->registerEngines([
-                '.volt'  => function ($view, $di) {
-                    /* @var \Phalcon\Di $di */
-                    $volt = new \Phalcon\Mvc\View\Engine\Volt($view, $di);
+        $view->registerEngines([
+            '.volt'  => function ($view, $di) {
+                /* @var \Phalcon\Di $di */
+                $volt = new \Phalcon\Mvc\View\Engine\Volt($view, $di);
 
-                    $volt->setOptions([
-                        'compiledPath'      => $di->getShared(Services::CONFIG)->view->compiledPath,
-                        'compiledSeparator' => '_'
-                    ]);
-                    $volt->getCompiler()->addExtension(new PhpFunctionExtension());
+                $volt->setOptions([
+                    'compiledPath'      => $di->getShared(Services::CONFIG)->view->compiledPath,
+                    'compiledSeparator' => '_'
+                ]);
+                $volt->getCompiler()->addExtension(new PhpFunctionExtension());
 
-                    return $volt;
-                },
-                '.phtml' => 'Phalcon\Mvc\View\Engine\Php'
-            ]);
+                return $volt;
+            },
+            '.phtml' => 'Phalcon\Mvc\View\Engine\Php'
+        ]);
 
-            return $view;
-        });
+        return $view;
+        // });
     }
 }
