@@ -2,6 +2,8 @@
 
 namespace Middleware;
 
+use Phalcon\Events\Event;
+
 /**
  * Class Middlewarize
  *
@@ -12,14 +14,14 @@ trait Middlewarize
     /**
      * Called on the initialization
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function init($event, $source, $data = null)
+    public function init(Event $event, $source, $data = null)
     {
         $this->view(__FUNCTION__, [$event, $source, $data]);
     }
@@ -27,14 +29,14 @@ trait Middlewarize
     /**
      * Called before the execution of handler
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function before($event, $source, $data = null)
+    public function before(Event $event, $source, $data = null)
     {
         $this->view(__FUNCTION__, [$event, $source, $data]);
     }
@@ -42,14 +44,14 @@ trait Middlewarize
     /**
      * Called after the execution of handler
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function after($event, $source, $data = null)
+    public function after(Event $event, $source, $data = null)
     {
         $this->view(__FUNCTION__, [$event, $source, $data]);
     }
@@ -57,14 +59,14 @@ trait Middlewarize
     /**
      * Called before the execution of handler
      *
-     * @param \Phalcon\Events\Event|mixed $event
+     * @param \Phalcon\Events\Event $event
      * @param \Phalcon\Dispatcher|mixed   $source
      * @param mixed|null                  $data
      *
      * @throws \Exception
      * @return bool
      */
-    public function finish($event, $source, $data = null)
+    public function finish(Event $event, $source, $data = null)
     {
         $this->view(__FUNCTION__, [$event, $source, $data]);
     }
